WIP Рендерит основной каркас страницы из презентера
