+++ src/view/most-commented.js
@@ -1,0 +1,17 @@
+import AbstractView from "./abstract.js";
+
+const createMostCommentedMoviesTemplate = () => {
+  return `<section class="films-list films-list--extra">
+  <h2 class="films-list__title">Most commented</h2>
+
+  <div class="films-list__container">
+
+  </div>
+</section>`;
+};
+
+export default class MostCommentedMoviesView extends AbstractView {
+  getTemplate() {
+    return createMostCommentedMoviesTemplate();
+  }
+}
