+++ src/view/top-rated.js
@@ -1,0 +1,15 @@
+import AbstractView from "./abstract.js";
+
+const createTopRatedMoviesTemplate = () => {
+  return `<section class="films-list films-list--extra">
+  <h2 class="films-list__title">Top rated</h2>
+  <div class="films-list__container">
+  </div>
+</section>`;
+};
+
+export default class TopRatedMoviesView extends AbstractView {
+  getTemplate() {
+    return createTopRatedMoviesTemplate();
+  }
+}
